Implement average values as doubles
Having average value calculations as unsigned integers was producing harsh
rounding errors for small valued quantities (like nprocs, nthreads). These are
now implemented as double precision values and get serialised as
numbers in JSON with decimal fractions.

There is a slight drawback that the decimal part is designed in nlohmann::json
to avoid precision loss, so sometimes a very high number of DPs are printed.
However, there doesn't seem to be any way to adjust this without manually
massaging the actual numerical values (there are various issues in
https://github.com/nlohmann/json where this is discussed).

+++ package/src/countmon.cpp
@@ -1,4 +1,4 @@
-// Copyright (C) CERN, 2018
+// Copyright (C) CERN, 2020
 
 #include "countmon.h"
 #include "utils.h"
@@ -49,7 +49,7 @@ void countmon::update_stats(const std::vector<pid_t>& pids) {
     if(count_stats[count_param] > count_peak_stats[count_param])
       count_peak_stats[count_param] = count_stats[count_param];
     count_total_stats[count_param] += count_stats[count_param];
-    count_average_stats[count_param] = count_total_stats[count_param] / iterations; 
+    count_average_stats[count_param] = double(count_total_stats[count_param]) / iterations; 
   }
 }
 
@@ -64,7 +64,7 @@ std::map<std::string, unsigned long long> const countmon::get_json_total_stats()
 }
 
 // An the averages 
-std::map<std::string, unsigned long long> const countmon::get_json_average_stats(
+std::map<std::string, double> const countmon::get_json_average_stats(
     unsigned long long elapsed_clock_ticks) {
   return count_average_stats;
 }
